+++ specs/outputhandler_routes_unknown_message_type_to_default_handler/psake-config.ps1
@@ -1,6 +1,0 @@
-$config.outputHandlers.Heading = $null
-$config.outputHandlers.Default = { Param($output) Write-Output "Default : $output" }
-$config.outputHandlers.Debug = $null
-$config.outputHandlers.Warning = $null
-$config.outputHandlers.Error = $null
-$config.outputHandlers.Success = $null